refactor: move recheckServerStatus into useSystemChecks, restore original return type
